requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
